+++ app/models/concerns/folio/audited.rb
@@ -1,0 +1,32 @@
+# frozen_string_literal: true
+
+module Folio::Audited
+  extend ActiveSupport::Concern
+
+  included do
+    attr_accessor :audit
+  end
+
+  module ClassMethods
+    def audited(opts = {})
+      super(opts)
+
+      # https://github.com/collectiveidea/audited/blob/master/lib/audited/auditor.rb#L125
+      # monkey patch: add related audit to revision
+      define_method(:revisions) do |from_version = 1|
+        targeted_audits = audits.from_version(from_version)
+
+        return [] unless targeted_audits
+
+        previous_attributes = reconstruct_attributes(audits - targeted_audits)
+
+        targeted_audits.map do |audit|
+          previous_attributes.merge!(audit.new_attributes)
+          previous_attributes[:audit_version] = audit.version
+          previous_attributes[:audit] = audit
+          revision_with(previous_attributes)
+        end
+      end
+    end
+  end
+end