+++ lib/folio.rb
@@ -35,6 +35,7 @@ require 'invisible_captcha'
 require 'rails-i18n'
 require 'aasm'
 require 'recaptcha'
+require 'audited'
 
 require 'folio/console/simple_form_inputs/date_time_input'