+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+
+## 2019-04-26
+### Added
+- `Folio::Audited` version control concern
+- `folio_pages_audited` to application config (enables version control for `Folio::Page`)
+
 ## 2019-04-18
 ### Added
 - added `devise_invitable` for `Folio::Account`